wip on refactor to reduce delta
