+++ constance/management/commands/constance.py
@@ -2,17 +2,29 @@
 
 from __future__ import unicode_literals
 
-from django import VERSION
 from django.core.exceptions import ValidationError
 from django.core.management import BaseCommand, CommandError
 from django.utils.translation import ugettext as _
 
+from ...admin import ConstanceForm
 from ...utils import get_values
-from ...forms import set_constance_value
-from ... import LazyConfig
+from ... import config
 
 
-config = LazyConfig()
+def _set_constance_value(raw_name, raw_value):
+    """
+    Parses and sets a Constance value
+    :param raw_name:
+    :param raw_value:
+    :return:
+    """
+
+    form = ConstanceForm(initial=get_values())
+
+    field = form.fields[raw_name]
+
+    value = field.clean(field.to_python(raw_value))
+    setattr(config, raw_name, value)
 
 
 class Command(BaseCommand):
@@ -41,7 +53,7 @@ class Command(BaseCommand):
                 raw_value = options.get('setting')[1]
 
                 try:
-                    set_constance_value(raw_name, raw_value)
+                    _set_constance_value(raw_name, raw_value)
                 except KeyError as e:
                     raise CommandError(raw_name + " is not defined in settings.CONSTANCE_CONFIG")
                 except ValidationError as e:
@@ -49,4 +61,4 @@ class Command(BaseCommand):
 
         elif options.get('list'):
             for k, v in get_values().items():
-                self.stdout.write("{}\t{}".format(k, v))
+                self.stdout.write("{}\t{}".format(k, v).encode('utf-8'))
